+++ CHANGELOG.md
@@ -18,7 +18,7 @@ but technically it is a breaking change to the Elasticsearch mappings.
   - Adds logging when messages fail to be added to queue
 
 - **CUMULUS-2577**
-  - Adds `POST` to `executions` endpoint to create an execution
+  - Adds `POST /executions` endpoint to create an execution
 
 ## [v9.4.0] 2021-08-16